// types.ts
export interface CompletedSet {
    checked: boolean;
    weight: string;
    reps?: string;
}

export interface Exercise {
  id: string; // Unique identifier for the exercise instance
  name: string;
  sets: string;
  reps: string;
  description:string;
  completedSets?: CompletedSet[];
  equipmentUsed?: string[]; // Equipment from user's selection needed for this exercise
}
// ...
export interface WorkoutPlan {
  title: string;
  duration: number;
  focus: string;
  exercises: Exercise[];
}